created modules,removed comments and console.log
